[EI-185] [web] refactor: Refactor IndicatorValue class and add ActualIndicatorValue subclass
