+++ web/__tests__/hooks/indicator/use-indicators-value-view-model.hook.test.tsx
@@ -5,7 +5,7 @@ import { useWorkspaceStore } from '@/app/store/stores/numerical-guidance/workspa
 import { resetAllStore } from '@/app/store/stores/reset-store';
 import { useLiveIndicatorsValueViewModel } from '@/app/business/hooks/indicator/use-live-indicators-value-view-model.hook';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
-import { IndicatorsValue } from '@/app/business/services/view-model/actual-indicators-value-view-model.service';
+import { ActualIndicatorsValue } from '@/app/business/services/view-model/indicator-value/actual-indicators-value-view-model.service';
 
 const wrapper = SWRProviderWithoutCache;
 
@@ -40,7 +40,7 @@ describe('useLiveIndicatorsValueViewModel', () => {
     await waitFor(() => expect(result.current.indicatorsValue).not.toBeUndefined());
 
     // then
-    expect(result.current.indicatorsValue).toBeInstanceOf(IndicatorsValue);
+    expect(result.current.indicatorsValue).toBeInstanceOf(ActualIndicatorsValue);
     expect(result.current.indicatorsValue?.length).toBe(1);
   });
 
@@ -74,7 +74,7 @@ describe('useLiveIndicatorsValueViewModel', () => {
     await waitFor(() => expect(result.current.indicatorsValue).not.toBeUndefined());
 
     // then
-    expect(result.current.indicatorsValue).toBeInstanceOf(IndicatorsValue);
+    expect(result.current.indicatorsValue).toBeInstanceOf(ActualIndicatorsValue);
     expect(result.current.indicatorsValue?.length).toBe(2);
   });
