+++ web/app/business/services/view-model/indicator-value/custom-forecast-indicator-value-view-model.service.ts
